Added an actual ScaleValue function and renamed existing ScaleVale funciton to UnScaleValue
The 1st ScaleValue function was actually an unscale function and I added an actual ScaleValuefunction. I also updated the SetServoPositions to  1st scale the joystick value from -1 to 1 into a 0 to 1 value. Then it UnScales the scaled value between each servos min and max value.

Also added some comments to the main loop to define what joystick is doing what